Try moving test into 'generic.py'
