+++ aiida/backends/djsite/db/subtests/generic.py
@@ -148,3 +148,14 @@ class TestDbExtrasDjango(AiidaTestCase):
         self.assertEquals(n1.dbnode.extras, new_attrs)
         # Also check that other nodes were not damaged
         self.assertEquals(n2.dbnode.extras, {'pippo2': [3, 4, 'b']})
+
+class TestBool(AiidaTestCase):
+    def test_bool_conversion(self):
+        from aiida.orm.data.base import Bool
+        for val in [True, False]:
+            self.assertEqual(val, bool(Bool(val)))
+
+    def test_int_conversion(self):
+        from aiida.orm.data.base import Bool
+        for val in [True, False]:
+            self.assertEqual(int(val), int(Bool(val)))
